refactor(core): extract metadata preview helper
